Move test files to __test__ folders
