+++ js/screens/__test__/Game.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Game from './Game';
+import Game from '../Game';
 
 import renderer from 'react-test-renderer';
